requirements.txt
apache-beam[gcp]
google-cloud-pubsub
faker
pandas
numpy